making wallet to work with EvenbetEngine from Gameserver.
changing Balance returning in responses from int to decimal.

+++ Platipus.Wallet.Api/Application/Requests/Wallets/Evenbet/EvenbetCreditRequest.cs
@@ -47,7 +47,7 @@ public sealed record EvenbetCreditRequest(
             var data = walletResult.Data;
 
             var response = new EvenbetCreditResponse(
-                (int)data!.Balance,
+                data!.Balance,
                 DateTimeOffset.Now.ToUnixTimeSeconds().ToString(),
                 data.Transaction.InternalId);
