+++ Platipus.Wallet.Api/Application/Responses/Evenbet/EvenbetDebitResponse.cs
@@ -3,6 +3,6 @@
 using Base;
 
 public sealed record EvenbetDebitResponse(
-    int Balance,
+    decimal Balance,
     string Timestamp,
     string TransactionId) : EvenbetCommonResponseWithTransaction(Balance, Timestamp, TransactionId);
